broken cam controllers, working custom faces

+++ src/main/java/com/example/MotorolaScienceCup/Battlezone/EnemyTank.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 
 public class EnemyTank extends Object3D{
 
-    private double rotation;
-    public double ENEMY_SPEED = 10;
+    private float rotation;
+    public float ENEMY_SPEED = 10;
 
 
 
-    public EnemyTank(double rotation, ArrayList<Vertex> points3D, ArrayList<Face> faces3D){
+    public EnemyTank(float rotation, ArrayList<Vertex> points3D, ArrayList<Face> faces3D){
         super(points3D, faces3D);
         this.rotation = rotation;
     }
